+++ compiler-core/src/ast/constant.rs
@@ -46,6 +46,17 @@ pub enum Constant<T, RecordTag> {
         record_constructor: Option<Box<ValueConstructor>>,
     },
 
+    RecordUpdate {
+        location: SrcSpan,
+        module: Option<(EcoString, SrcSpan)>,
+        name: EcoString,
+        record: Box<Self>,
+        arguments: Vec<ConstantRecordUpdateArg<Self>>,
+        tag: RecordTag,
+        type_: T,
+        field_map: Option<FieldMap>,
+    },
+
     BitArray {
         location: SrcSpan,
         segments: Vec<BitArraySegment<Self, T>>,
@@ -85,6 +96,7 @@ impl TypedConstant {
             }
             Constant::List { type_, .. }
             | Constant::Record { type_, .. }
+            | Constant::RecordUpdate { type_, .. }
             | Constant::Var { type_, .. }
             | Constant::Invalid { type_, .. } => type_.clone(),
         }
@@ -107,6 +119,16 @@ impl TypedConstant {
             Constant::Record { arguments, .. } => arguments
                 .iter()
                 .find_map(|argument| argument.find_node(byte_index))
+                .unwrap_or(Located::Constant(self)),
+            Constant::RecordUpdate {
+                record, arguments, ..
+            } => record
+                .find_node(byte_index)
+                .or_else(|| {
+                    arguments
+                        .iter()
+                        .find_map(|arg| arg.value.find_node(byte_index))
+                })
                 .unwrap_or(Located::Constant(self)),
             Constant::BitArray { segments, .. } => segments
                 .iter()
@@ -139,6 +161,7 @@ impl TypedConstant {
             } => value_constructor
                 .as_ref()
                 .map(|constructor| constructor.definition_location()),
+            Constant::RecordUpdate { .. } => None,
         }
     }
 
@@ -160,6 +183,17 @@ impl TypedConstant {
                 .iter()
                 .map(|argument| argument.value.referenced_variables())
                 .fold(im::hashset![], im::HashSet::union),
+
+            Constant::RecordUpdate {
+                record, arguments, ..
+            } => {
+                let record_vars = record.referenced_variables();
+                let arg_vars = arguments
+                    .iter()
+                    .map(|arg| arg.value.referenced_variables())
+                    .fold(im::hashset![], im::HashSet::union);
+                record_vars.union(arg_vars)
+            }
 
             Constant::BitArray { segments, .. } => segments
                 .iter()
@@ -194,6 +228,7 @@ impl<A, B> Constant<A, B> {
             | Constant::Tuple { location, .. }
             | Constant::String { location, .. }
             | Constant::Record { location, .. }
+            | Constant::RecordUpdate { location, .. }
             | Constant::BitArray { location, .. }
             | Constant::Var { location, .. }
             | Constant::Invalid { location, .. }
@@ -212,6 +247,7 @@ impl<A, B> Constant<A, B> {
             Constant::Tuple { .. }
             | Constant::List { .. }
             | Constant::Record { .. }
+            | Constant::RecordUpdate { .. }
             | Constant::BitArray { .. }
             | Constant::StringConcatenation { .. }
             | Constant::Invalid { .. } => false,
@@ -232,5 +268,28 @@ impl<A, B> bit_array::GetLiteralValue for Constant<A, B> {
         } else {
             None
         }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct ConstantRecordUpdateArg<Constant> {
+    pub label: EcoString,
+    pub location: SrcSpan,
+    pub value: Constant,
+}
+
+impl<Constant> ConstantRecordUpdateArg<Constant> {
+    #[must_use]
+    pub fn uses_label_shorthand(&self) -> bool
+    where
+        Constant: HasLocation,
+    {
+        self.value.location() == self.location
+    }
+}
+
+impl<Constant> HasLocation for ConstantRecordUpdateArg<Constant> {
+    fn location(&self) -> SrcSpan {
+        self.location
     }
 }