+++ CHANGELOG.md
@@ -51,6 +51,14 @@
   containing scientific notation or trailing zeros (i.e. `100` and `1e2`).
   ([ptdewey](https://github.com/ptdewey))
 
+- Record update syntax now works with spreading records in type constructors
+  for const values:
+  ```gleam
+  const a = Foo(1, 2)
+  const b = Foo(..a, 3)
+  ```
+  ([Adi Salimgereyev](https://github.com/abs0luty))
+
 ### Build tool
 
 - The help text displayed by `gleam dev --help`, `gleam test --help`, and